Split `CoverArt.cs` into multiple parts
`CoverArt.Internals.cs` contains all the internals,
`CoverArt.EndPoints.cs` contains the public API for the various API
endpoints, and `CoverArt.cs` contains the constructors, properties
and `HttpClient`/`IDispose` handling.

This also makes the class `sealed`.
