+++ public-api/MetaBrainz.MusicBrainz.CoverArt.net6.0.cs.md
@@ -12,7 +12,7 @@
 ### Type: CoverArt
 
 ```cs
-public class CoverArt : System.IDisposable {
+public sealed class CoverArt : System.IDisposable {
 
   public const int MaxImageSize = 536870912;
